refactor: replace validation from utils to config and template. replace template validation error to templates initialization.
